Add encryption and decryption logic with RSA key management
Implemented an RSA-based Encryptor to handle message encryption and decryption,
including key pair generation and signature verification. Added unit tests
to validate encryption integrity, decryption accuracy, and compatibility between clients.
